requirements.txt
# Quant MVP - Requirements
# Python 3.11+

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
duckdb>=0.9.0

# Data sources
yfinance>=0.2.28
requests>=2.31.0
requests-cache>=1.1.0

# Feature engineering
ta-lib>=0.4.28

# Modeling
lightgbm>=4.1.0
catboost>=1.2.0
scikit-learn>=1.3.0
shap>=0.42.0

# Validation & backtesting
vectorbt>=0.25.0

# Execution
alpaca-trade-api>=3.0.0

# Scheduling & ops
APScheduler>=3.10.0
python-telegram-bot>=20.0

# Experiment tracking
mlflow>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
